b2260: introduce ST platform b2260 (stih410 SoC, ARMv7 based).
b2260 is based on a STiH410 SoC itself based on an ARMv7 Cortex-A9
core with a PL310 level2 external cache. This change introduces the
b2260 with means to boot 2 possible secure firmware acting as a BL32
stage: the trusted firmware sp_min and the OP-TEE.

This change does not provide a full-features platform. Only the basics
are proposed here. Other features as SMP/PSCI will be proposed later on.

Change-Id: I310320d9cfc5f768fe75c2f948ad930297347472
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>