drivers/scmi-msg: driver for parsing scmi messages
This change introduces drivers that gets an input SCMI message and
parse it to call platform handlers. SCMI here allows secure world to
expose resources to the non-secure world or other subsystem of the
platform.

scmi_process_message() is called with a message context as argument.
The message context define where to find the input message payload,
from secure memory, and where to write the response message payload.
When it returns, the message when read and response written.

scmi_process_message() calls the SCMI protocol driver according to
the protocol ID in the message. The SCMI protocol driver will call
defined platform handlers according to the message content.

This change introduces only the SCMI base protocol as defined in
SCMI specification v2.0 [1]. Not all the messages defined
in the specification are supported.

[1] https://developer.arm.com/architectures/system-architectures/software-standards/scmi

Change-Id: I0b3ebc6e7de62e28fb063bbb5a62fb8d0e9d9fbd
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>