common: import speculation_barrier.h
Import speculation_barrier.h from [1] based on [2]. This file implements
a set of wrapper macros for the __builtin_load_no_speculate builtin
function detailed at [3]. This builtin function defines a speculation
barrier, which can be used to limit the conditions under which a
value which has been loaded can be used under speculative execution.

The header file provided here allows a migration path to using the
builtin function for users who are unable to immediately upgrade to a
compiler which supports the builtin.

(Description taken from the README.md file of the speculation_barrier
project by Arm on GitHub [4])

Link: [1] https://raw.githubusercontent.com/ARM-software/speculation-barrier/80e65cd830b321ff676941c2fab773da450e6dae/speculation_barrier.h
Link: [2] https://github.com/OP-TEE/optee_os commit 6aca2c8e640e ("libutils: import speculation_barrier.h")
Link: [3] https://www.arm.com/security-update
Link: [4] https://github.com/ARM-software/speculation-barrier

Change-Id: I1b4e006a0766ff42913aad3e254df7cb6a4cc4e1
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>