drivers: stm32mp1 clocks: support shifted clock selector bit masks
The current implementation optimizes memory consumes by gateable
clock table by storing bit mask and bit shift with 1 byte each.
The issue is that register selector bit masks above the 7th LSBit
cannot be stored.

This change uses the shift info to shift the mask bbefore it is used,
allowing clock selctor register bit fields to be spread on the 32 bits
of the register as long as the mask fits in 8 contiguous bit at most.

This change is needed to add the RTC clock to the gateable clocks table.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>